+++ src/main/java/de/samply/frontend/dto/DtoFactory.java
@@ -1,27 +1,66 @@
 package de.samply.frontend.dto;
 
-import de.samply.db.model.ProjectDocument;
+import de.samply.project.state.ProjectBridgeheadState;
 import jakarta.validation.constraints.NotNull;
-import org.springframework.stereotype.Service;
 
-@Service
+import java.time.Instant;
+
 public class DtoFactory {
 
-    public Project convert (@NotNull de.samply.db.model.Project project){
-        return new Project();
-        //TODO
+    public static Project convert(@NotNull de.samply.db.model.Project project) {
+        return new Project(
+                project.getCode(),
+                project.getCreatorEmail(),
+                project.getCreatedAt(),
+                project.getExpiresAt(),
+                project.getArchivedAt(),
+                project.getModifiedAt(),
+                project.getState(),
+                project.getType(),
+                project.getQuery().getQuery(),
+                project.getQuery().getHumanReadable(),
+                project.getQuery().getQueryFormat(),
+                project.getQuery().getOutputFormat(),
+                project.getQuery().getTemplateId(),
+                project.getQuery().getLabel(),
+                project.getQuery().getDescription(),
+                project.getQuery().getExplorerUrl(),
+                project.getQuery().getContext()
+        );
     }
 
-    public Notification convert (@NotNull de.samply.db.model.Notification notification){
-        return new Notification();
-        //TODO
+    public static Notification convert(@NotNull de.samply.db.model.Notification notification) {
+        return new Notification(
+                notification.getEmail(),
+                notification.getTimestamp(),
+                notification.getProject().getCode(),
+                notification.getBridgehead(),
+                notification.getOperationType(),
+                notification.getDetails(),
+                notification.getError()
+        );
     }
 
-    public OtherDocument convert(@NotNull ProjectDocument projectDocument){
-        return new OtherDocument();
-        //TODO
+    public static ProjectDocument convert(@NotNull de.samply.db.model.ProjectDocument projectDocument) {
+        return new ProjectDocument(
+                projectDocument.getProject().getCode(),
+                projectDocument.getOriginalFilename(),
+                projectDocument.getUrl(),
+                projectDocument.getCreatedAt(),
+                projectDocument.getBridgehead(),
+                projectDocument.getCreatorEmail(),
+                projectDocument.getLabel(),
+                projectDocument.getDocumentType()
+        );
     }
 
-
+    public static ProjectBridgehead convert(@NotNull de.samply.db.model.ProjectBridgehead projectBridgehead) {
+        return new ProjectBridgehead(
+                projectBridgehead.getProject().getCode(),
+                projectBridgehead.getBridgehead(),
+                projectBridgehead.getState(),
+                projectBridgehead.getModifiedAt()
+        );
+    }
 
 }