+++ src/main/java/de/samply/frontend/dto/Notification.java
@@ -1,4 +1,16 @@
 package de.samply.frontend.dto;
 
-public record Notification() {
+import de.samply.notification.OperationType;
+
+import java.time.Instant;
+
+public record Notification(
+        String email,
+        Instant timestamp,
+        String projectCode,
+        String bridgehead,
+        OperationType operationType,
+        String details,
+        String error
+) {
 }