+++ src/main/java/de/samply/frontend/dto/OtherDocument.java
@@ -1,4 +1,0 @@
-package de.samply.frontend.dto;
-
-public class OtherDocument {
-}