+++ src/main/java/de/samply/frontend/dto/Project.java
@@ -1,4 +1,30 @@
 package de.samply.frontend.dto;
 
-public record Project() {
+import de.samply.project.ProjectType;
+import de.samply.project.state.ProjectState;
+import de.samply.query.OutputFormat;
+import de.samply.query.QueryFormat;
+
+import java.time.Instant;
+import java.time.LocalDate;
+
+public record Project(
+        String code,
+        String creatorEmail,
+        Instant createdAt,
+        LocalDate expiresAt,
+        Instant archivedAt,
+        Instant modifiedAt,
+        ProjectState state,
+        ProjectType type,
+        String query,
+        String humanReadable,
+        QueryFormat queryFormat,
+        OutputFormat outputFormat,
+        String templateId,
+        String label,
+        String description,
+        String explorerUrl,
+        String queryContext
+) {
 }