+++ src/main/java/de/samply/frontend/dto/ProjectBridgehead.java
@@ -1,4 +1,13 @@
 package de.samply.frontend.dto;
 
-public record ProjectBridgehead() {
+import de.samply.project.state.ProjectBridgeheadState;
+
+import java.time.Instant;
+
+public record ProjectBridgehead(
+        String projectCode,
+        String bridgehead,
+        ProjectBridgeheadState state,
+        Instant modifiedAt
+) {
 }